+++ podonos/core/file.py
@@ -8,13 +8,16 @@ class File:
     _tags: List[str]
     _script: Optional[str]
 
-    def __init__(self, path: str, model_tag: str, tags: List[str] = [], script: Optional[str] = None) -> None:
+    def __init__(self, path: str, model_tag: str, tags: List[str] = [], script: Optional[str] = None,
+                 is_ref: bool = False) -> None:
         """
         Args:
             path: Path to the file to evaluate. Required.
             model_tag: String that represents the model or group. Required.
             tags: A list of string for file. Optional.
             script: Script of the input audio in text. Optional.
+            is_ref: True if this file is to be a reference for an evaluation type that requires a reference.
+                    Optiona. Default is False.
         """
         log.check_ne(path, "")
         log.check_ne(model_tag, "")
@@ -22,6 +25,7 @@ class File:
         self._model_tag = model_tag
         self._tags = tags
         self._script = script
+        self._is_ref = is_ref
 
     @property
     def path(self) -> str:
@@ -38,3 +42,7 @@ class File:
     @property
     def script(self) -> Optional[str]:
         return self._script
+
+    @property
+    def is_ref(self) -> Optional[bool]:
+        return self._is_ref