+++ tests/test_double_stimuli_evaluator.py
@@ -10,11 +10,12 @@ from podonos.core.evaluation import Evaluation
 from podonos.core.file import File
 from podonos.errors.error import NotSupportedError
 from podonos.evaluators.double_stimuli_evaluator import DoubleStimuliEvaluator
-
+from tests.test_audio import TESTDATA_SPEECH_TWO_CH1_WAV, TESTDATA_SPEECH_TWO_CH2_WAV
 
 class MockDoubleStimuliEvaluator(DoubleStimuliEvaluator):
-    def __init__(self, supported_evaluation_type: List[EvalType], api_client=Mock(spec=APIClient), eval_config: Optional[EvalConfig] = None):
-        super().__init__(supported_evaluation_type, api_client, eval_config)
+    def __init__(self, supported_evaluation_types: List[EvalType],
+                 api_client=Mock(spec=APIClient), eval_config: Optional[EvalConfig] = None):
+        super().__init__(supported_evaluation_types, api_client, eval_config)
 
     def _create_evaluation(self) -> Evaluation:
         evaluation_config = {
@@ -35,60 +36,48 @@ class TestDoubleStimuliEvaluator:
         self.api_client = Mock(spec=APIClient)
         self.eval_config = EvalConfig(type="SMOS")
         self.evaluator = MockDoubleStimuliEvaluator(
-            supported_evaluation_type=[EvalType.SMOS], api_client=self.api_client, eval_config=self.eval_config
+            supported_evaluation_types=[EvalType.SMOS], api_client=self.api_client, eval_config=self.eval_config
         )
 
-    def test_add_file_pair_without_init(self):
+    def test_add_files_without_init(self):
         self.evaluator._initialized = False
         target_audio_config = File(path="generated.wav", model_tag="model_target", tags=["target"],
                                    script="This is the target audio file")
-        ref_audio_config = File(path="original.wav", model_tag="model_ref", tags=["reference"], script=None)
+        ref_audio_config = File(path="original.wav", model_tag="model_ref", tags=["reference"], script=None,
+                                is_ref=True)
 
         with pytest.raises(ValueError) as excinfo:
-            self.evaluator.add_file_pair(target=target_audio_config, ref=ref_audio_config)
+            self.evaluator.add_files(file0=target_audio_config, file1=ref_audio_config)
 
-        assert "Try to add_file_pair once the evaluator is closed." in str(excinfo.value)
+        assert "evaluator is not initialized" in str(excinfo.value)
 
     def test_add_file_not_supported(self):
         with pytest.raises(NotSupportedError) as excinfo:
             self.evaluator.add_file(File(path="1.wav", model_tag="model1", tags=["test"]))
-        assert "The 'add_file' is only supported in these evaluation types: {'NMOS', 'QMOS', 'P808'}" in str(excinfo.value)
+        assert "only supported" in str(excinfo.value)
 
-    def test_add_file_pair_unsupported_eval_type(self):
+    def test_add_files_unsupported_eval_type(self):
         self.evaluator._initialized = True
 
         target = File(path="target.wav", model_tag="model_target", tags=["generated"])
-        ref = File(path="ref.wav", model_tag="model_ref", tags=["human"])
+        ref = File(path="ref.wav", model_tag="model_ref", tags=["human"], is_ref=True)
 
         with pytest.raises(ValueError) as excinfo:
-            self.evaluator.add_file_pair(target=target, ref=ref)
+            self.evaluator.add_files(file0=target, file1=ref)
 
-        assert "The add_file_pair function is used for 'CMOS', 'DMOS'" in str(excinfo.value)
+        assert "Unsupported" in str(excinfo.value)
 
-    def test_add_file_set_unsupported_eval_type(self):
+    def test_add_files_missing_ref(self):
         self.evaluator._initialized = True
         self.eval_config._eval_type = EvalType.CMOS
 
-        file0 = File(path="file1.wav", model_tag="model0", tags=["file1"], script="We are the world")
-        file1 = File(path="file2.wav", model_tag="model1", tags=["file2"])
+        file0 = File(path="file0.wav", model_tag="model0", tags=["file0"], script="We are the world")
+        file1 = File(path="file1.wav", model_tag="model1", tags=["file1"])
 
         with pytest.raises(ValueError) as excinfo:
-            self.evaluator.add_file_set(file0=file0, file1=file1)
+            self.evaluator.add_files(file0=file0, file1=file1)
 
-        assert "The add_file_set function is used for 'SMOS', 'PREF'" in str(excinfo.value)
-
-    def test_close_upsupported_eval_type(self):
-        self.evaluator._initialized = True
-        self.eval_config._eval_type = EvalType.PREF
-
-        file0 = File(path="file1.wav", model_tag="model0", tags=["file1"])
-        file1 = File(path="file2.wav", model_tag="model1", tags=["file2"])
-        self.evaluator.add_file_set(file0=file0, file1=file1)
-
-        with pytest.raises(ValueError) as excinfo:
-            self.evaluator.close()
-
-        assert "Not supported evaluation type" in str(excinfo.value)
+        assert "must be set" in str(excinfo.value)
 
     def test_generate_random_group_name(self):
         group_name = self.evaluator._generate_random_group_name()