+++ tests/test_single_stimulus_evaluator.py
@@ -13,8 +13,9 @@ from podonos.evaluators.single_stimulus_evaluator import SingleStimulusEvaluator
 
 
 class MockDoubleStimuliEvaluator(SingleStimulusEvaluator):
-    def __init__(self, supported_evaluation_type: List[EvalType], api_client=Mock(spec=APIClient), eval_config: Optional[EvalConfig] = None):
-        super().__init__(supported_evaluation_type, api_client, eval_config)
+    def __init__(self, supported_evaluation_types: List[EvalType], api_client=Mock(spec=APIClient),
+                 eval_config: Optional[EvalConfig] = None):
+        super().__init__(supported_evaluation_types, api_client, eval_config)
 
     def _create_evaluation(self) -> Evaluation:
         evaluation_config = {
@@ -35,7 +36,8 @@ class TestSingleStimulusEvaluator:
         self.api_client = Mock(spec=APIClient)
         self.eval_config = EvalConfig(type="NMOS")
         self.evaluator = MockDoubleStimuliEvaluator(
-            supported_evaluation_type=[EvalType.NMOS, EvalType.QMOS, EvalType.P808], api_client=self.api_client, eval_config=self.eval_config
+            supported_evaluation_types=[EvalType.NMOS, EvalType.QMOS, EvalType.P808],
+            api_client=self.api_client, eval_config=self.eval_config
         )
 
     def test_add_file_before_initialized(self):
@@ -44,16 +46,9 @@ class TestSingleStimulusEvaluator:
             self.evaluator.add_file(file=File(path="./male.wav", model_tag="model1", tags=["male"]))
         assert "Try to add file once the evaluator is closed." in str(excinfo.value)
 
-    def test_add_file_pair_not_supported(self):
+    def test_add_files_not_supported(self):
         target_audio_config = File(path="target.wav", model_tag="model2", tags=["target"])
         ref_audio_config = File(path="ref.wav", model_tag="model1", tags=["ref"])
         with pytest.raises(NotSupportedError) as excinfo:
-            self.evaluator.add_file_pair(target=target_audio_config, ref=ref_audio_config)
-        assert "The 'add_file_pair' is only supported in these evaluation types: {'CMOS', 'DMOS'}" in str(excinfo.value)
-
-    def test_add_file_set_not_supported(self):
-        file0 = File(path="original.wav", model_tag="original", tags=["human"], script="Dog is very cute")
-        file1 = File(path="generated.wav", model_tag="model1", tags=["generated"], script="Dog is very cute")
-        with pytest.raises(NotSupportedError) as excinfo:
-            self.evaluator.add_file_set(file0=file0, file1=file1)
-        assert "The 'add_file_set' is only supported in these evaluation types: {'SMOS', 'PREF'}" in str(excinfo.value)
+            self.evaluator.add_files(target=target_audio_config, ref=ref_audio_config)
+        assert "only supported" in str(excinfo.value)